Move new settings page to default settings page & update stylesheet
